Shell-escape program name and args
Currently the program and args are just joined and passed in; vterm inserts this
string after `exec` in its sh command. This causes quotes and other shell
characters to be misinterpreted and also strips empty arguments. By quoting each
of the args before joining them, the intended command line is executed.